fix(gui): circular import
