+++ api-editor/gui/src/features/packageData/model/PythonFromImport.ts
@@ -1,5 +1,5 @@
 import { Optional } from '../../../common/util/types';
-import { PythonFromImportJson } from './PythonPackageBuilder';
+import {PythonFromImportJson} from "./APIJsonData";
 
 export class PythonFromImport {
     constructor(readonly module: string, readonly declaration: string, readonly alias: Optional<string> = null) {}
