+++ api-editor/gui/src/features/packageData/model/PythonImport.ts
@@ -1,5 +1,5 @@
 import { Optional } from '../../../common/util/types';
-import { PythonImportJson } from './PythonPackageBuilder';
+import {PythonImportJson} from "./APIJsonData";
 
 export class PythonImport {
     constructor(readonly module: string, readonly alias: Optional<string> = null) {}
