+++ api-editor/gui/src/features/packageData/model/PythonPackage.ts
@@ -4,8 +4,8 @@ import { Optional } from '../../../common/util/types';
 import { PythonClass } from './PythonClass';
 import { PythonFunction } from './PythonFunction';
 import { PythonParameter } from './PythonParameter';
-import { PythonPackageJson } from './PythonPackageBuilder';
 import { EXPECTED_API_SCHEMA_VERSION } from '../apiSlice';
+import {PythonPackageJson} from "./APIJsonData";
 
 interface PythonPackageShallowCopy {
     distribution?: string;
