+++ api-editor/gui/src/features/packageData/model/PythonPackageBuilder.ts
@@ -1,4 +1,3 @@
-import { Optional } from '../../../common/util/types';
 import { PythonClass } from './PythonClass';
 import { PythonFromImport } from './PythonFromImport';
 import { PythonFunction } from './PythonFunction';
@@ -9,16 +8,16 @@ import { PythonParameter, PythonParameterAssignment } from './PythonParameter';
 import { PythonResult } from './PythonResult';
 import { PythonDeclaration } from './PythonDeclaration';
 import { EXPECTED_API_SCHEMA_VERSION } from '../apiSlice';
-
-export interface PythonPackageJson {
-    schemaVersion?: number;
-    distribution: string;
-    package: string;
-    version: string;
-    modules: PythonModuleJson[];
-    classes: PythonClassJson[];
-    functions: PythonFunctionJson[];
-}
+import {
+    PythonClassJson,
+    PythonFromImportJson,
+    PythonFunctionJson,
+    PythonImportJson,
+    PythonModuleJson,
+    PythonPackageJson,
+    PythonParameterJson,
+    PythonResultJson,
+} from './APIJsonData';
 
 export const parsePythonPackageJson = function (packageJson: PythonPackageJson): PythonPackage | null {
     if ((packageJson.schemaVersion ?? 1) !== EXPECTED_API_SCHEMA_VERSION) {
@@ -47,15 +46,6 @@ export const parsePythonPackageJson = function (packageJson: PythonPackageJson):
         idToDeclaration,
     );
 };
-
-export interface PythonModuleJson {
-    id: string;
-    name: string;
-    imports: PythonImportJson[];
-    from_imports: PythonFromImportJson[];
-    classes: string[];
-    functions: string[];
-}
 
 const parsePythonModuleJson = function (
     moduleJson: PythonModuleJson,
@@ -100,37 +90,13 @@ const parsePythonModuleJson = function (
     return result;
 };
 
-export interface PythonImportJson {
-    module: string;
-    alias: Optional<string>;
-}
-
 const parsePythonImportJson = function (importJson: PythonImportJson): PythonImport {
     return new PythonImport(importJson.module, importJson.alias);
 };
 
-export interface PythonFromImportJson {
-    module: string;
-    declaration: string;
-    alias: Optional<string>;
-}
-
 const parsePythonFromImportJson = function (fromImportJson: PythonFromImportJson): PythonFromImport {
     return new PythonFromImport(fromImportJson.module, fromImportJson.declaration, fromImportJson.alias);
 };
-
-export interface PythonClassJson {
-    id: string;
-    name: string;
-    qname: string;
-    decorators: string[];
-    superclasses: string[];
-    methods: string[];
-    is_public: boolean;
-    reexported_by: string[];
-    description: Optional<string>;
-    docstring: Optional<string>;
-}
 
 const parsePythonClassJson = function (
     classJson: PythonClassJson,
@@ -158,19 +124,6 @@ const parsePythonClassJson = function (
     return result;
 };
 
-export interface PythonFunctionJson {
-    id: string;
-    name: string;
-    qname: string;
-    decorators: string[];
-    parameters: PythonParameterJson[];
-    results: PythonResultJson[];
-    is_public: boolean;
-    reexported_by: string[];
-    description: Optional<string>;
-    docstring: Optional<string>;
-}
-
 const parsePythonFunctionJson = function (
     functionJson: PythonFunctionJson,
     idToDeclaration: Map<string, PythonDeclaration>,
@@ -190,20 +143,6 @@ const parsePythonFunctionJson = function (
     idToDeclaration.set(functionJson.id, result);
     return result;
 };
-
-export interface PythonParameterJson {
-    id: string;
-    name: string;
-    qname: string;
-    default_value: Optional<string>;
-    assigned_by: 'IMPLICIT' | 'POSITION_ONLY' | 'POSITION_OR_NAME' | 'NAME_ONLY';
-    is_public: boolean;
-    docstring: {
-        type: Optional<string>;
-        description: Optional<string>;
-    };
-    type: object; // TODO parse type
-}
 
 const parsePythonParameterJson = function (
     parameterJson: PythonParameterJson,
@@ -238,14 +177,6 @@ const parsePythonParameterAssignment = function (
             return PythonParameterAssignment.NAME_ONLY;
     }
 };
-
-export interface PythonResultJson {
-    name: string;
-    docstring: {
-        type: Optional<string>;
-        description: Optional<string>;
-    };
-}
 
 const parsePythonResultJson = function (resultJson: PythonResultJson): PythonResult {
     return new PythonResult(resultJson.name, resultJson.docstring.type ?? '', resultJson.docstring.description ?? '');
